+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning][].
 - Add ``dotools_py.utility.set_path`` to create and return a normalize path.
 - Add ``dotools_py.utility.to_toml`` to save information of the current environment to TOML file for sharing.
 - Add utility functions for plotting
+- Improvement of the performance of ``do.get.mean_expr`` and ``do.get.expr``
 
 ### Bug Fixes
 - Minor bug fixes for several functions